Version 2.4.0 Update

+++ EdgeWare/panic.bat
@@ -1,3 +1,9 @@
 @echo off
 color 5d
-taskkill /F /IM pythonw.exe
+pushd %~dp0
+if exist hid_time.dat (
+    echo Nice try cutie~
+    pause
+) else (
+    taskkill /F /IM pythonw.exe
+)
